Factorize code between WebApp tests files
