+++ web/src/test/java/org/ensembl/gti/genesearch/rest/EndpointTests.java
@@ -56,45 +56,10 @@ import static org.junit.Assert.*;
 @RunWith(SpringJUnit4ClassRunner.class)
 @SpringBootTest(classes = Application.class,
         webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@SpringBootApplication(exclude = {SecurityAutoConfiguration.class})
+@SpringBootApplication
 @TestExecutionListeners(DependencyInjectionTestExecutionListener.class)
-public class EndpointTests {
+public class EndpointTests extends WebAppTests {
 
-
-    @Value("${local.server.port}")
-    int port;
-
-    private String GENES_FETCH = "genes/fetch";
-    private String GENES_QUERY = "genes/query";
-    private String GENOMES_FETCH = "genomes/fetch";
-    private String GENOMES_QUERY = "genomes/query";
-    private String GENOMES_SELECT = "genomes/select";
-    private String TRANSCRIPTS_FETCH = "transcripts/fetch";
-    private String TRANSCRIPTS_QUERY = "transcripts/query";
-    private String INFO = "genes/info";
-
-    private static Logger log = LoggerFactory.getLogger(EndpointTests.class);
-    private static ESTestClient esTestClient;
-
-    static ESSearch geneSearch;
-    static ESSearch genomeSearch;
-
-    @BeforeClass
-    public static void setUp() throws IOException {
-        // create our ES test server once only
-        log.info("Setting up ");
-        esTestClient = new ESTestClient();
-        log.info("Reading documents");
-        String geneJson = DataUtils.readGzipResource("/nanoarchaeum_equitans_kin4_m_genes.json.gz");
-        String genomeJson = DataUtils.readGzipResource("/genomes.json.gz");
-        log.info("Creating test index");
-        esTestClient.indexTestDocs(geneJson, ESSearch.GENES_INDEX, ESSearch.GENE_ESTYPE);
-        esTestClient.indexTestDocs(genomeJson, ESSearch.GENOMES_INDEX, ESSearch.GENOME_ESTYPE);
-
-    }
-
-    @Autowired
-    EndpointSearchProvider provider;
 
     static final TypeReference<Map<String, Object>> MAP_REF = new TypeReference<Map<String, Object>>() {
     };
@@ -104,17 +69,6 @@ public class EndpointTests {
     };
 
     private RestTemplate restTemplate = new TestRestTemplate().getRestTemplate();
-
-    private String getServiceUrl(String uri) {
-        return "http://localhost:" + port + "/api/" + uri;
-    }
-
-    @Before
-    public void injectSearch() {
-        // ensure we always use our test instances
-        log.info("Inject search " + esTestClient.getClient());
-        provider.setESClient(esTestClient.getClient());
-    }
 
     @Test
     public void testQueryGetEndpoint() {
